# services/quiz/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API квизов, вопросов, ответов и пользователей quiz-service.\nПубличные ответы не содержат is_correct, а пользователи — хешей паролей.",
        "title": "QuizGo CRUD API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/answers": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Доступно только автору квиза. Если is_correct не передан, используется false. Тело запроса должно содержать один JSON-объект без неизвестных полей.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "answers"
                ],
                "summary": "Создать ответ",
                "parameters": [
                    {
                        "description": "Новый ответ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/quizhandler.CreateAnswerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ответ создан; тело ответа пустое"
                    },
                    "400": {
                        "description": "Некорректный JSON, ID вопроса или пустой текст",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Недействительный токен или пользователь не является автором квиза",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Вопрос не найден",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/answers/{id}": {
            "get": {
                "description": "Публичное представление ответа без признака is_correct.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "answers"
                ],
                "summary": "Получить ответ",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID ответа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quizhandler.PublicAnswerResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID ответа",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Ответ не найден",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Доступно только автору квиза.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "answers"
                ],
                "summary": "Удалить ответ",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID ответа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ответ удалён; тело ответа пустое"
                    },
                    "400": {
                        "description": "Некорректный ID ответа",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Недействительный токен или пользователь не является автором квиза",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Ответ не найден",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/answers/{id}/author": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает ответ с признаком is_correct. Доступно только автору квиза.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "answers"
                ],
                "summary": "Получить ответ для автора",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID ответа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quizhandler.AuthorAnswerResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID ответа",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Недействительный токен или пользователь не является автором квиза",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Ответ не найден",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "quizhandler.AuthorAnswerResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_correct": {
                    "type": "boolean",
                    "example": true
                },
                "question_id": {
                    "type": "integer",
                    "example": 1
                },
                "text_content": {
                    "type": "string",
                    "example": "Париж"
                }
            }
        },
...
        "quizhandler.CreateAnswerRequest": {
            "type": "object",
            "required": [
                "question_id",
                "text_content"
            ],
            "properties": {
                "is_correct": {
                    "type": "boolean",
                    "default": false,
                    "example": true
                },
                "question_id": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "text_content": {
                    "description": "Текст не должен быть пустым после удаления пробелов по краям.",
                    "type": "string",
                    "minLength": 1,
                    "example": "Париж"
                }
            }
        },
...
        "quizhandler.PublicAnswerResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "question_id": {
                    "type": "integer",
                    "example": 1
                },
                "text_content": {
                    "type": "string",
                    "example": "Париж"
                }
            }
        },
...
        "respond.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите Bearer \u003cJWT\u003e. Токен можно получить через POST /auth/login с JSON {\"username\":\"alice\",\"password\":\"secret123\"}. Поле token уже содержит префикс Bearer. Регистрация: POST /auth/register с теми же полями.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}